Split Ktor module database
