+++ ztor/src/main/kotlin/com/zenmo/ztor/Application.kt
@@ -6,13 +6,11 @@ import com.zenmo.orm.connectToPostgres
 import com.zenmo.orm.createSchema
 import com.zenmo.orm.echoSchemaSql
 import com.zenmo.ztor.plugins.*
-import com.zenmo.zummon.companysurvey.surveysFromJson
 import com.zenmo.zummon.companysurvey.surveysToJson
 import io.ktor.server.application.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
 import kotlin.system.exitProcess
-import kotlin.uuid.Uuid
 
 fun main(args: Array<String>) {
     if (args.count() == 1) {
@@ -69,8 +67,10 @@ fun Application.module() {
     configureMonitoring()
     configureSerialization()
     configureAuthentication()
-    val db = configureDatabases()
+    val db = connectToPostgres()
+    configureUserEndpoints(db)
     configureSurveys(db)
+    configureProjects(db)
     configureRouting()
     configureStatusPages()
     configureUpload(db)
